Refactor the QObject QWebChannel and make the API reusable.
It might becme a very common use case of the QWebChannel QML plugin.
Thus it should be as simple as possible for third party consumers to
setup a QWebChannel for QObject publishing.

The new API basically moves the QtMetaObjectPublisher along with
the JavaScript marshalling to the qwebchannl/src folder.

The updated qtobject example shows how this new API can be used.
Furthermore note how it is now trivially possible to register
multiple objects, which was not easily possible before.

Some notes on the applied refactoring:

- qobject.js contains the JavaScript QObject binding and was
  refactored to support multiple objects.
- the MetaObjectPublisher contains a new handleRequest function
  which handles the QML-side of the QObject binding. This is
  implemented in QML, while the other book keeping and esp. the
  classInfoForObject is still handled in C++ via the
  QtMetaObjectPublisher class (which is registered as
  MetaObjectPublisherPrivate and used by MetaObjectPublisher)

Change-Id: Id45121bb654447e095bf8a8062d0c8edf9dcb018