Fix long-polling issue of dropped events.
This was especially noticeable when multiple timers where running.
In such cases only the very first signal would get submitted, then
the poll socket got closed. Consecutive calls to broadcast got lost.

Now we save the data passed to QWebChannel::broadcast and submit
it all in one go whenever a poll request comes in.

This fixes the timer issue and all signals get submitted. In the long
term this should all be properly rewritten using a WebSocket server
on the QML side and a WebSocket client on the HTML side.

Change-Id: I3b26f073978cf14b9bd045ec28b0279128c3c9e6