Cleanup sources, mostly by removing QtCreator generated bloat.
This removes a lot of obsolete files and simplifies the build system
of the examples.

Furthermore, the examples can now be run without running make install
first. It reuses the same import path as the test does.

Note that the examples are not installable anymore now though. If this
is required, it can be added again.

Change-Id: Ic7ff80f734b035a03fb1a11a2df492c97298ceff