Add an initial test suite for the MetaObjectPublisher
This tests the functionality of publishing a plain QtObject from
QML to the HTML client.

It tests property binding, i.e. reading and writing of an objects
property on the client side, as well as change notification tracking.

Furthermore a server-side method is invoked from the client and
signal submission from the server to the client is tested.

Change-Id: I62e544cddf4483b57535a9bc1e05a36105ec6622
Reviewed-by: Zeno Albisser <zeno.albisser@digia.com>