Optimize QtMetaObjectPublisher::classInfoForObject.
This includes two optimizations which have been forgotten to be
upstreamed before:

a) For the common case of property notify signals named "...Changed",
the string name is now not send anymore to the client. Instead, these
cases are special-cased and constructed on the fly. This drastically
reduces the size of Qt.init responses and property update messages.

b) Furthermore, do not list signals as methods, further reducing the
size of Qt.init response messages.

Together this shows a noticeable reduce of CPU instructions in the
benchmarks as recorded with perf:

benchmark_classInfo: down ~10%
benchmark_initializeClients: down ~2%
benchmark_propertyUpdates: down ~1%

Change-Id: I01e59f5c1dceedb893f7a3e3e127acb493baaa7f
Reviewed-by: Michael Bruning <michael.bruning@digia.com>
Reviewed-by: Pierre Rossi <pierre.rossi@gmail.com>