Test QObject wrapping, when factory-style methods are called.
An HTML client can trigger the creation of new objects, or a published
object on the server side might return other objects.

These are then wrapped on the fly and can be used like the other
objects. Note though that the HTML client can call deleteLater on
these objects. Also, it does not yet work to wrap objects on the fly
in signal arguments or property values.

Change-Id: I92aa8a3e52f42d5325dd0771bbf9e2ae213e88f9
Reviewed-by: Arvid Nilsson <anilsson@blackberry.com>
Reviewed-by: Pierre Rossi <pierre.rossi@gmail.com>