Optimize and cleanup code by moving WebChannel.qml logic into C++.
By handling more logic on the C++ side, such as serializing the JSON
data to QByteArray, we can save up to 6% in the propertyUpdates
benchmark.

Furthermore, this allows for some more code cleanup and obsoletes the
WebChannel.qml file.

Also, it is a first step towards removing the QML dependency and
making it optional alltogether.

Change-Id: Id610b5f2652da4a7ad867aef576fabcc40d3d92c
Reviewed-by: Pierre Rossi <pierre.rossi@gmail.com>