Restructure sources and assimilate to Qt module structure.
This module can hopefully be done in time for 5.3. This commit changes
the source structure and QMake files to adapt to typical Qt modules.

With this in place, we can now use QT += webchannel in qmake files to
link against the pure Qt/C++ QtWebChannel library.

The QML plugin is separated from it and can be loaded optionally, if
the quick module could be found. Also added is now a qmlplugindump
for tooling integration. Note that the Qt.labs namespace is removed.

The test file structure is also adapted to how its done in the
QtDeclarative module.

Note that this setup apparently does not support to run tests without
running make install first.

Change-Id: I1c15d72e7ab5f525d5a6f651f4e965ef86bc17bd
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>