Add standalone C++/Qt example which opens HTML client in a browser.
This example shows how to use the (currently quite ugly) raw C++ API
to setup a webchannel without using QML at all. The HTML client is then
handled by the users default browser.

The example itself shows a simple chat between the HTML client and the
C++/Qt server, with a line edit for input and a text edit showing the
chat history.

Change-Id: I8baf14efb9d0c5f5880d99710cf6317fe9b887b9