Optimize: Use property-indices instead of names for notifications.
This reduces the traffic as the indices are usually much smaller than
the property names.

As such, the benchPropertyUpdates gets a speed boost of about 9% (or
10ms vs. 11ms). As we need to transmit the index during initialization
that degrades its performance slightly by ca. 4% (13ms vs. 12.5ms).

Considering that the initialization only takes place once whereas
the property updates potentially often, this is a good tradeoff.

Change-Id: If7df3e360f1528b7d7aa26c63ce851363ae9fd6a