Fix regression in handling of var-emitting signals from QML.
Qt 5.3 propagates var-arguments of signals as QJSValue instead of as a
QVariant. This then fails to be serialized to QJson, failing our unit
tests.

Now, the basic QJSValue types are handled manually to make the tests
pass again. Note though that more complicated JS types such as arrays
or objects will be propagated as strings. This should be improved
eventually by transforming these into the analogous QJson types.

Change-Id: I730c595eee214ebe3d1f83009cd5605f66407f55