Mark QtWebKit as optional module dependency.
The current CI system fails to build the webchannel when
QtWebKit is included in sync.profile, so we ignore it for now
and adapt the QMake build system to only include the declarative
tests when webkit is found.

As discussed with tronical, sifalt and sahumada, the qtqa scripts
will need to be adapted to cope with optional dependencies of non-qt5
modules.

Change-Id: Id89b763ef2697e9e72eb4064c150971b13ebccc3
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>