Refactor response handling in order to send data only to target client.
Before, the response was sent to all clients in a broad-cast and had
to be filtered on the client-side. This required additional client
identification data to be added to all requests and responses.

Now, we keep track of the transport and transport-internal client and
only send the response to that client.

Change-Id: Ia1ef5e031b0058222083d352a8aa28a7d566a6ca