Refactor code to use QWebChannelAbstractTransport and QtWebSockets.
This is a quite big changeset, but neccessary to get the roadmap
implemented that was discussed at QtCS.

With this patchset landed, the QWebChannel does not depend on
QtWebKit anymore, not even for the tests. Rather, we will introduce
the dependency in the other way (i.e. QtWebKit will optionally use
QtWebChannel if available).

For the pure Qt/C++ use-case, we ship a utility implementation of
a QWebChannelAbstractTransport that uses a QWebSocket for the
server-client communication. This way, we can get rid of the custom
WebSocket implementation.

The tests are refactored to run the qwebchannel.js code directly
inside QML. Integration tests for QtWebKit/QtWebEngine as well
as examples will be added to these repositories.

Change-Id: Icc1c1c5918ec46e31d5070937c14c4ca25a3e2d6