Wrap QmlWebSocket and QQuickWebViewExperimental in QMPI object.
Instead of copying the interface to external Qt modules, we now wrap
the known useful objects directly in QmlWebChannel on the fly.

This is easily achieved using QObject introspection as we can rely on
the public QML API of both, QmlWebSocket and QQuickWebViewExperimental.

Personally, I think this is the better approach compared to copying
the interface around and implementing it in other external modules which
should not need to be aware of QWebChannel.

This will be easily extensible to future navigator.qt-like API in the
QML WebEngineView object.

Change-Id: I268c6af1e625cf9e6308ac943d10e457a2c9c5d7