Only depend optionally on QtWebSockets, and only use it in the example.
The utility QWebChannelAbstractTransport implementation based on the
QtWebSocket has no big value. Instead, it would pull in the QtWebSocket
link-time dependency into QtWebKit/QtWebEngine, which is not desired.
Considering that the WebSocket usecase is minor, and only few people
will ever use it, we agreed that having the code in the example alone
is enough.

Change-Id: Ica038329a1d684f33e805fc296e9dff71b1446ba