WIP: Use JavaScript prototype inheritance.
I'm not sure whether this is cleaner or better compared to what we had
before. Also, the same should be done for the JS QObject class if this
turns out to be better.

Also, this gave me another idea of how to potentially improve the code
in the future: We could send QObject class information seperately from
the instance data. That way, if we publish multiple objects of the
same type, we could share the same prototype code on the client side.

Anyhow, I'm up for discussion here. If it's not worth it, I'm also
fine with discarding this changeset.

Change-Id: I32f313c765658e231e21b24386eff389bd806229