Added some more examples
Implemented a C++ Server, providing a chatservice through a simple
QObject using QWebChannel and also both a QML and HTML Client to
interact with the server.
The examples show how to use properties, signals and methods,
separated in server and clients.

This reverts commit e8c280a15f7822fa5a563388615313858fd90848.
The integration issues w.r.t. unconditional usage of the
websockets module have been resolved.

Change-Id: I4a07646dcd5e6050d3bd6b9bd1c5b07dbc7d7d69
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>