Fix static builds and building of tests
For static linkage we need to explicitly reference the plugin factory class, to
prevent the linker from discarding the symbol. The Qt build system takes care of
generating the necessary code for referencing the symbol and linking that reference
into the application - see also qt.prf. For the build system to do this automatically,
it needs to know the class name, which is stored in the qmldir file.

This patch adds the missing class name. The first failure for this was when
building the tests/auto/qml test, which specifies an IMPORTPATH that causes the
qml import scanner to find the erroneous qmldir file and error out (see
qt.prf). However otherwise this error would've appeared when trying to link an
app on say iOS for example.

Change-Id: Ia28307593e651d038ff1daaea7597f6eb5e40eb1
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>