Example: Modify JavaScript part of the HTML
The baseUrl that is used to connect to a WebSocket server is derived
from the query parameters set to the URL. These parameters are ignored
by the QDesktopServices::openUrl implementations specific to Windows
and OS X. Ubuntu uses the default implementation that retains
the query parameters. This means the browser instance will fail to open
the local file URL on Ubuntu because it includes the query parameters,
so query parameters should never be set.

Moreover, the example creates a QWebSocketServer instance, which is
available at ws://localhost:12345, so the it is safe to hardcode the
address in the JavaScript. The cleaner approach would be to use a
URLHandler as described in
http://doc.qt.io/qt-5/qdesktopservices.html#url-handlers, but that
complicates the example.

Change-Id: I5b5df2b7b816ce0bbfb16a85c036ed379616f04a
Task-number: QTBUG-46541
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>