Example: Add code to copy the JS file from the resource system
The JS file lives in the src/webchannel directory, and there
is a QMake magic in place to copy it to the build directory. But
this mechanism fails when the example is run from the Qt binary
pkg, which does not include the sources. The JS file must be copied
to the build directory either manually or programmatically to run
the example.

Change-Id: Ib56d9348a8bf1a599e2db5235e0545cd7a8f3bb1
Task-number: QTBUG-46541
Reviewed-by: Topi Reiniö <topi.reinio@theqtcompany.com>
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>