Convert QObjects in QVariantMaps
QObjects that are present in an array are already converted to
an object identifier.  This does the same for variant maps, which
end up as ECMAScript objects.

This allows QObjects put into a QVariantMap to be properly
deserialized.  For example, if a property is declared as such:

Q_PROPERTY(QVariantMap propName READ propName CONSTANT)

And propName is:

QVariantMap propName() const {
  QVariantMap map;
  map.insert("theProperty", QVariant::fromValue(someQObject));
  return map;
}

The "theProperty" property will now properly refer to the object.

Change-Id: I3c6e71b860f6825a31eb337aeffa55302287c8ff
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>