Unwrap property updates in case they are QObjects
When an object is initialized, QObject properties are properly
unwrapped.  However, before this patch, updates made to QObject
properties were not unwrapped and thus could not be used from
ECMAScript.

For example, given a property defined as such:

Q_PROPERTY(QObject* theProp READ theProp NOTIFY thePropChanged)

The property "theProp" is correctly unwrapped when the parent
object is created.  However, it would not have been unwrapped
when "thePropChanged" is emitted.  This patch corrects that.

Change-Id: If8f67560d9fb2a4bd909d2ab30305ceab30f8d31
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>