Pack QVariants into args when QVariant is expected
When a method expects an argument of type QVariant, the argument should
actually contain a QVariant as opposed to the converted QVariant value.
If the converted value is given, which was the logic before this patch,
then some values may not be properly converted back to a QVariant when
the method is actually invoked (null for example).

For example, if a method signature is as follows:

  Q_INVOKABLE void test(QVariant value);

And it is called from ECMAScript via WebChannel as follows:

  obj.test(null)

Then the argument would have been created as type 'std::nullptr_t'
and passed into the method.  This results in a QVariant with a
random type, which could cause a warning or even a segmentation fault
depending on what the type ends up as.

Instead, this patch creates the argument as a 'QVariant' which
results in the proper type when the method is invoked.

Change-Id: I3f75048da1c0f32707f6f162844de74157f0b535
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>