Restore compatibility with custom QVariantList/QVariantMap converters
QVariant::toValue<QVariantList> does not honor explicit custom converters
for a sequentially iterable container to a QVariantList. This breaks
compatibility with Qt 5.12.3 and before, where the old code path using
QVariant::toList() used the custom converter. Do the same for
QVariantMap.

This patch restores the compatibility with old code. The real fix will
target the QVariantList converter code in QtCore - this is just a hotfix
within Qt WebChannel for now.

Task-number: QTBUG-80751
Change-Id: Ic70c6a353aad43ddbaefbc6626a0af87bd0d024f
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>