Fix documentation to use commonly working syntax for property changes
The "on" prefix and capitalization of the first letter in the change
signal name is only done in QML, not in JavaScript. Furthermore, not
all notify signals carry the new value in their signature.

Adapt the example code to not show the "on" prefix and also use the
more reliable approach to read a property value directly.

Task-number: QTBUG-81782
Change-Id: I9c25e48c2d7abcdabbca88526f62c3d79eb952f8
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>