Fix infinite recursion when wrapping a self-contained object twice
The previous unit test for self-contained objects only wrapped the test
object once. After wrapping, a different code path is taken which still
exhibited infinite recursion. This patch addresses both the unit test
and the infinite recursion.
To fix the problem, a boolean in the ObjectInfo struct is toggled to
indicate whether the object in question is currently being wrapped. If
that is the case, the recursing code path is skipped.

[ChangeLog][General] Fixed infinite recursion when dealing with self
contained objects.

Fixes: QTBUG-84007
Change-Id: Ie0898fb5f28cec91587897835ff937672d60f2a1
Reviewed-by: Kai Koehne <kai.koehne@qt.io>
(adapted from commit 0451ef836415c93a6beb68a315a25c6ab27f44fa)
Reviewed-by: Paul Wicking <paul.wicking@qt.io>