Update commercial license headers
Updated header.COMM to the files in tqtc-qtwebchannel. Examples,
tests or documentation files are not updated.

The commercial license header may contain
some additional lines so that its line count
equals with the earlier license header. Reason for
this is that some autotests use hard coded line numbers
and a change in the line count causes failures
in tests.

Task-number: QTQAINFRA-4219
Change-Id: I5e919619b7123243b848efa639b7bbe5e3deae56
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>