Fix up the recently added test for QProperty/QWebChannel interaction
The recently added code was not quite conformant to the coding style and
did not properly use Q_OBJECT_BINDABLE_PROPERTY together with a notify
signal. This commit fixes both.

The commit also adds a test to verify that changes to a QProperty are
actually transmitted over QWebChannel. At the moment, this relies on
notify signals. A follow-up patch will use onValueChanged() to propagate
changes of properties without a NOTIFY signal.

Task-number: QTBUG-91757
Change-Id: Ia5a323cba02dc72df774beeba445d40140d7af7f
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>