Update commercial license headers
Updated header.COMM to the files in tqtc-qtwebchannel.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4963
Change-Id: Iec9fed7a8c016fd5933067b1528f79affd8d2c4b
Reviewed-by: Antti Kokko <antti.kokko@qt.io>