tst_webchannel: port from raw and shared pointers to unique_ptr
Avoids the need for qDeleteAll() in the case of raw pointers. Avoids
overhead for non-shared objects in the shared pointer case.

Since unique_ptr is a move-only type, use a container (std::vector)
and loop constuct (ranged for loop) that can handle it.

This removes the only use of QSharedPointer in this module.

Change-Id: Ic3fe90403eb3f28f1e2e370a3648bc522880a10d
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Arno Rehn <a.rehn@menlosystems.com>