Make chatclient-qml example executable
This example used to have to be run using the qml application, now it is
a stand-alone application. Some changes are made to make it work in Dark
Mode and to improve the user experience. The documentation is changed to
reflect the changes and it is slightly expanded.

Task-number: QTBUG-113000
Change-Id: I65f0db74e5bbf4cb917b4da980a54dd85c14a954
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit e89f78f21c16e707ebd8ba1a18c05f10ae982f09)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>